# backend/swagger/docs/swagger.yaml
definitions:
...
  iteminfo.ExtendedFileInfo:
    properties:
      checksums:
        additionalProperties:
          type: string
        description: checksums for the file
        type: object
      content:
        description: text content of a file, if requested
        type: string
      files:
        description: files in the directory with optional metadata
        items:
          $ref: '#/definitions/iteminfo.ExtendedItemInfo'
        type: array
      folders:
        description: folders in the directory
        items:
          $ref: '#/definitions/iteminfo.ItemInfo'
        type: array
      hasPreview:
        description: whether the file has a thumbnail preview
        type: boolean
      hash:
        description: hash for the file -- used for sharing
        type: string
      hidden:
        description: whether the file is hidden
        type: boolean
      isShared:
        description: whether the file or folder is shared
        type: boolean
      metadata:
        allOf:
        - $ref: '#/definitions/iteminfo.MediaMetadata'
        description: media metadata for audio/video files (includes duration)
      modified:
        description: modification time
        type: string
      name:
        description: name of the file
        type: string
      onlyOfficeId:
        description: id for onlyoffice files
        type: string
      path:
        description: path scoped to the associated index
        type: string
      pinnedItems:
        description: pinned item names in this directory listing
        items:
          type: string
        type: array
      size:
        description: length in bytes for regular files
        type: integer
      source:
        description: associated index source for the file
        type: string
      subtitles:
        description: subtitles for video files
        items:
          $ref: '#/definitions/utils.SubtitleTrack'
        type: array
      token:
        description: token for the file -- used for sharing
        type: string
      type:
        description: type of the file, either "directory" or a file mimetype
        type: string
      viewToken:
        description: opaque token for inline viewing via /resources/view or /media/stream
        type: string
    type: object
  iteminfo.ExtendedItemInfo:
    properties:
      hasPreview:
        description: whether the file has a thumbnail preview
        type: boolean
      hidden:
        description: whether the file is hidden
        type: boolean
      isShared:
        description: whether the file or folder is shared
        type: boolean
      metadata:
        allOf:
        - $ref: '#/definitions/iteminfo.MediaMetadata'
        description: optional media metadata (audio/video only)
      modified:
        description: modification time
        type: string
      name:
        description: name of the file
        type: string
      size:
        description: length in bytes for regular files
        type: integer
      type:
        description: type of the file, either "directory" or a file mimetype
        type: string
      viewToken:
        description: opaque token for inline viewing via /resources/view or /media/stream
        type: string
    type: object
...
  iteminfo.ItemInfo:
    properties:
      hasPreview:
        description: whether the file has a thumbnail preview
        type: boolean
      hidden:
        description: whether the file is hidden
        type: boolean
      isShared:
        description: whether the file or folder is shared
        type: boolean
      modified:
        description: modification time
        type: string
      name:
        description: name of the file
        type: string
      size:
        description: length in bytes for regular files
        type: integer
      type:
        description: type of the file, either "directory" or a file mimetype
        type: string
    type: object
  iteminfo.Lyric:
    properties:
      text:
        type: string
      timestamp:
        description: milliseconds
        type: integer
    type: object
  iteminfo.MediaMetadata:
    properties:
      album:
        description: album name
        type: string
      albumArt:
        description: album art image data (automatically base64-encoded in JSON)
        items:
          type: integer
        type: array
      artist:
        description: track artist
        type: string
      audioCodec:
        description: audio codec name (e.g. aac)
        type: string
      container:
        description: container/format name (e.g. mov,mp4,m4a,3gp,3g2,mj2)
        type: string
      duration:
        description: duration in seconds
        type: integer
      genre:
        description: music/video genre
        type: string
      hasLyrics:
        description: checks if lyrics are available without parse them
        type: boolean
      lyrics:
        description: lyrics (from embedded tags or .lrc files)
        items:
          $ref: '#/definitions/iteminfo.Lyric'
        type: array
      title:
        description: track/video title
        type: string
      track:
        description: track number
        type: integer
      videoCodec:
        description: video codec name (e.g. h264)
        type: string
      year:
        description: release year
        type: integer
    type: object
...
  utils.SubtitleTrack:
    properties:
      codec:
        description: codec name for embedded subtitles
        type: string
      embedded:
        description: true for embedded subtitles, false for external files
        type: boolean
      index:
        description: stream index for embedded subtitles (nil for external)
        type: integer
      language:
        description: language code
        type: string
      name:
        description: filename for external, or descriptive name for embedded
        type: string
      title:
        description: title/description
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/media/lyrics:
    get:
      consumes:
      - application/json
      description: Returns parsed lyrics with optional timestamps from embedded tags
        or sidecar .lrc files.
      parameters:
      - description: Path to the directory or file
        in: query
        name: path
        required: true
        type: string
      - description: Source name
        in: query
        name: source
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Lyrics array
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get lyrics for an audio file
      tags:
      - Media
  /api/media/metadata:
    get:
      consumes:
      - application/json
      description: Same ExtendedFileInfo as resources GET with metadata=true (typically
        used for directories).
      parameters:
      - description: Path to the directory or file
        in: query
        name: path
        required: true
        type: string
      - description: Source name
        in: query
        name: source
        required: true
        type: string
      - description: When true, include embedded album art bytes in audio metadata
        in: query
        name: albumArt
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/iteminfo.ExtendedFileInfo'
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Directory with media metadata
      tags:
      - Media
...
  /api/media/subtitles:
    get:
      consumes:
      - application/json
      description: Returns raw subtitle content from external files or embedded streams
      parameters:
      - description: Index path to the video file
        in: query
        name: path
        required: true
        type: string
      - description: Source name for the desired source
        in: query
        name: source
        required: true
        type: string
      - description: Subtitle track name (filename for external, descriptive name
          for embedded)
        in: query
        name: name
        required: true
        type: string
      - description: Whether this is an embedded stream (true) or external file (false),
          defaults to false
        in: query
        name: embedded
        type: boolean
      produces:
      - text/plain
      responses:
        "200":
          description: Raw subtitle content in original format
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Resource not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get subtitle content
      tags:
      - Media
...
  /public/api/media/lyrics:
    get:
      parameters:
      - description: Share hash
        in: query
        name: hash
        required: true
        type: string
      - description: Path within the share
        in: query
        name: path
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Lyrics array
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get lyrics for an audio file (public share)
      tags:
      - Media
  /public/api/media/metadata:
    get:
      parameters:
      - description: Share hash
        in: query
        name: hash
        required: true
        type: string
      - description: Path within the share
        in: query
        name: path
        type: string
      - description: When true, include embedded album art bytes in audio metadata
          (heavier)
        in: query
        name: albumArt
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/iteminfo.ExtendedFileInfo'
      summary: Directory with media metadata (public share)
      tags:
      - Media
...
swagger: "2.0"